api_ocr/requirements.txt
fastapi
uvicorn
pillow
pytesseract
python-multipart